==> analyses/molecular-subtyping-HGG/00-HGG-select-pathology-dx.Rmd <==
---
title: "Select pathology diagnoses for inclusion"
output: 
  html_notebook:
    toc: TRUE
    toc_float: TRUE
author: Jaclyn Taroni for ALSF CCDL
date: 2020
---

## Background

Originally, we subtyped tumors in this module if the specimen satisfied one of the following criteria:

1. A defining lesion was identified in the SNV consensus file (H3 K28M or G35R/V)
2. The `short_histology` was `HGAT`.

In an upcoming release, `integrated_diagnosis`, which can be updated as the result of subtyping, will be used to populate the `short_histology` column (see [#748](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/748)).
Thus, molecular subtyping modules need to be upstream of `short_histology` and use the `pathology_diagnosis` and `pathology_free_text_diagnosis` fields.
This change for this module is tracked in [#754](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/754).

Filtering on the basis of `short_histology == HGAT` is more straightforward than using the pathology diagnosis fields, so we include this notebook to put together the terms in `pathology_diagnosis` and `pathology_free_text_diagnosis`.

We will use the 2016 WHO Classification as our guide ([Louis et al. _Acta Neuropathol._ doi: 10.1007/s00401-016-1545-1](10.1007/s00401-016-1545-1)) and take a look at the current version of the histology file (`release-v17-20200908`).

## Set up

```{r}
library(tidyverse)
```

### Directories and files

We're going to tie this to a specific release.

```{r}
data_dir <- file.path("..", "..", "data", "release_003_008", "release_003_008")
histologies_file <- file.path(data_dir, "pbta-histologies.tsv")
```

We're going to save the pathology diagnosis information we'll use to generate the subset files in a directory `hgg-subset`.

```{r}
output_dir <- "hgg-subset"
output_file <- file.path(output_dir, 
                         "hgg_subtyping_path_dx_strings.json")
```

## Read in data

```{r}
histologies_df <- read_tsv(histologies_file)
```

## Explore the pathology diagnoses

### `short_histology == HGAT`

In the current histologies file, if we filter based on `short_histology` as we did originally, what is in the pathology diagnosis fields?
Note that some of the `short_histology` values will have been altered based on earlier subtyping efforts.
(That's why we're doing this!)

```{r}
histologies_df %>% 
  filter(short_histology == "HGAT") %>%
  count(pathology_diagnosis) %>%
  arrange(desc(n))
```

For the most part, this is as we would expect given the 2016 WHO classifications. 
In an initial round of subtyping, PNET specimens were reclassified (see [this comment on #609](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/609#issuecomment-602821376)).
We should not and will not include PNET samples in the criteria used for detect samples for subtyping from pathology diagnosis fields. 
Instead, these samples that were reclassified earlier shouldbe included downstream on the basis of defining lesions.

Although some of these terms appear to be part of a defined vocabulary, there are others like `High Grade Glial Neoplasma` that were not subject to the same harmonization.
These are likely from the completed PNOC trial (see [#754 (comment)] (https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/754#issuecomment-697004412)).
Let's take a look if we filter by `cohort == "PNOC003"`.


```{r}
histologies_df %>%
  filter(cohort == "PNOC003",
         # Filter out normal WGS rows
         sample_type != "Normal") %>%
  count(pathology_diagnosis) %>%
  arrange(desc(n))
```

As anticipated, these are all indicative samples that should be included for subtyping but they are not harmonized.

Let's take a look at the free text field when filtering with `short_histology == HGAT`.

```{r}
histologies_df %>% 
  filter(short_histology == "HGAT") %>%
  group_by(pathology_free_text_diagnosis) %>%
  tally() %>%
  arrange(desc(n))
```

As we might expect from a free text field, this is even less uniform.

## Pathology diagnosis strings for inclusion

For the CBTTC samples, the `pathology_diagnosis` fields are harmonized, so we can use the terms below to look for exact matches.

```{r}
exact_path_dx<- c(
  "High-grade glioma/astrocytoma (WHO grade III/IV)",
  "Brainstem glioma- Diffuse intrinsic pontine glioma",
  "Gliomatosis Cerebri"
)
```

And all samples from the PNOC003 trial should be included.

Let's take a look at a first attempt using these terms as described above.

```{r}
filtered_on_dx_df <- histologies_df %>%
  filter(pathology_diagnosis %in% exact_path_dx | 
            cohort == "PNOC003",
         # Exclude normal samples when filtering on cohort
         sample_type != "Normal") %>%
  select(Kids_First_Biospecimen_ID, 
         sample_id, 
         Kids_First_Participant_ID,
         pathology_diagnosis,
         pathology_free_text_diagnosis,
         integrated_diagnosis, 
         short_histology)

filtered_on_dx_df
```

Let's tally the values in `pathology_diagnosis` in this data frame.

```{r}
filtered_on_dx_df %>%
  count(pathology_diagnosis) %>%
  arrange(desc(n))
```

We are not including any samples with pathology diagnoses outside of what we should include for subtyping.

### Save the strings we'll use downstream

Create a list with the strings we'll use for inclusion.

```{r}
terms_list <- list(exact_path_dx = exact_path_dx)
```

Save this list as JSON.

```{r}
writeLines(jsonlite::prettify(jsonlite::toJSON(terms_list)), output_file)
```

## Session Info

```{r}
sessionInfo()
```
