==> analyses/molecular-subtyping-HGG/05-HGG-molecular-subtyping-fusion.Rmd <==
---
title: "High-Grade Glioma Molecular Subtyping - Fusions"
output: 
  html_notebook:
    toc: TRUE
    toc_float: TRUE
author: Chante Bethell and Jaclyn Taroni for ALSF CCDL
date: 2020
---

This notebook prepares putative oncogenic data for the purpose of subtyping HGG samples 
([`AlexsLemonade/OpenPBTA-analysis#249`](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/249)).

## Usage

This notebook is intended to be run via the command line from the top directory
of the repository as follows:

```
Rscript -e "rmarkdown::render('analyses/molecular-subtyping-HGG/05-HGG-molecular-subtyping-fusion.Rmd', clean = TRUE)"
```

## Set up

### Libraries and Functions

```{r}
library(tidyverse)
```

### Directories

```{r}
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
# File path to results directory
input_dir <-
  file.path(root_dir, "analyses", "molecular-subtyping-HGG", "hgg-subset")
# File path to results directory
results_dir <-
  file.path(root_dir, "analyses", "molecular-subtyping-HGG", "results")
if (!dir.exists(results_dir)) {
  dir.create(results_dir)
}
```

### Read in Files

The fusion file was generated by filtering `pbta-fusion-putative-oncogenic.tsv` file to samples that met the criteria for HGG molecular subtyping.

```{r}
fusion_df <- read_tsv(file.path(input_dir, "hgg_fusion.tsv"))
```

### Output File

```{r}
output_file <- file.path(results_dir, "HGG_cleaned_fusion.tsv")
```

## What fusions are we interested in?

There are two genes of interest according to [`AlexsLemonade/OpenPBTA-analysis#249`](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/249).
Specifically:

* _FGFR1_ fusions should be mutually exclusive of H3 K28 mutations 
* _NTRK_ fusions are co-occurring with H3 G35 mutations

In the original issue, there is no mention of specific fusion partners or orientations.
Thus, we will look at _any_ instances of fusions that contain _FGFR1_ or _NTRK_.

## Filter and Wrangle Fusion Data

```{r}
FGFR1_fusions <- fusion_df %>%
  filter(str_detect(FusionName, "FGFR1"))

dim(FGFR1_fusions)
```

There is a single _FGFR1_ fusion present in the filtered and prioritized fusion table for samples that were included in the HGG subset file.
Recall that samples in the HGG subset file either had a defining lesion (H3 K28 or H3 G35 mutation) _or_ were labeled `HGAT` in the `short_histology` column of `pbta-histologies.tsv`.

_FGFR1_ fusions are _mutually exclusive_ so this result is not necessarily unexpected.

```{r}
NTRK_fusions <- fusion_df %>%
  filter(str_detect(FusionName, "NTRK"))

NTRK_fusions
```

_NTRK_ refers to [a _family_ of receptor kinases](https://www.biooncology.com/pathways/cancer-tumor-targets/ntrk/ntrk-oncogenesis.html) and not an individual gene symbol.
To retain information about which gene we're talking about we will include the full fusion name in the table when we present this information.

```{r}
ntrk_summary_df <- NTRK_fusions %>%
  select(Sample, FusionName) %>%
  group_by(Sample) %>%
  summarize(NTRK_fusions = paste(sort(unique(FusionName)), collapse = ", "))
```

Add back in the samples without fusions.

```{r}
subset_bsids <- unique(fusion_df$Sample)
ntrk_summary_df <- ntrk_summary_df %>%
  # Add in the samples -- the NTRK_fusion column will get filled with NA
  bind_rows(data.frame(Sample = setdiff(subset_bsids, 
                                        ntrk_summary_df$Sample))) %>%
  # Replace NA with "None" because the data are not missing - we simply do not
  # have evidence for an NTRK fusion in that sample
  replace(is.na(.), "None")

head(ntrk_summary_df, n = 10)
```

No samples had evidence of _FGFR1_ fusions - we'll add a column to indicate that
and write to file.

```{r}
ntrk_summary_df %>%
  mutate(FGFR1_fusions = if_else(
    Sample %in% FGFR1_fusions$Sample,
    "FGFR1--TACC1",
     "None"
  )) %>%
  dplyr::arrange(Sample) %>%
  write_tsv(output_file)
```

## Session Info

```{r}
sessionInfo()
```
